+++ packages/components-native/src/index.ts
@@ -1,2 +1,4 @@
 export * from "./Icon";
 export * from "./Divider";
+export * from "./Typography";
+export * from "./Text";