+++ packages/components/src/FormField/FormField.css.d.ts
@@ -13,6 +13,7 @@ declare const styles: {
   readonly "inputWrapper": string;
   readonly "childrenWrapper": string;
   readonly "input": string;
+  readonly "select": string;
   readonly "label": string;
   readonly "postfix": string;
   readonly "affixIcon": string;