+++ packages/components-native/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.2.1...@jobber/components-native@0.3.0) (2023-05-12)
+
+
+### Features
+
+* **components-native:** Import Text to components-native ([#1248](https://github.com/GetJobber/atlantis/issues/1248)) ([9c38749](https://github.com/GetJobber/atlantis/commit/9c387498e61c67b80d4704b8438dfc01cde6e853))
+
+
+
+
+
+## [0.2.1](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.2.0...@jobber/components-native@0.2.1) (2023-05-12)
+
+**Note:** Version bump only for package @jobber/components-native
+
+
+
+
+
+# [0.2.0](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.1.4...@jobber/components-native@0.2.0) (2023-05-12)
+
+
+### Features
+
+* **components-native:** Import Typography to components-native ([#1235](https://github.com/GetJobber/atlantis/issues/1235)) ([109fbae](https://github.com/GetJobber/atlantis/commit/109fbae26ce9a10d3fe5f0709fc86bfb60d05ca5))
+
+
+
+
+
+## [0.1.4](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.1.3...@jobber/components-native@0.1.4) (2023-05-12)
+
+**Note:** Version bump only for package @jobber/components-native
+
+
+
+
+
+## [0.1.3](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.1.2...@jobber/components-native@0.1.3) (2023-05-11)
+
+**Note:** Version bump only for package @jobber/components-native
+
+
+
+
+
 ## [0.1.2](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.1.1...@jobber/components-native@0.1.2) (2023-05-10)
 
 **Note:** Version bump only for package @jobber/components-native